LPD-100207 Share the Top Pages card layout in shared components
